Changed tests directory structure
