+++ test/Emitter/AbstractSapiEmitterTest.php
@@ -10,7 +10,7 @@
 
 declare(strict_types=1);
 
-namespace BitFrame\Test\Integration;
+namespace BitFrame\Test\Emitter;
 
 use PHPUnit\Framework\TestCase;
 use Psr\Http\Message\{ServerRequestInterface, ResponseInterface};
