+++ test/Http/ContentNegotiatorTest.php
@@ -8,10 +8,9 @@
  * @license   https://bitframephp.com/about/license MIT License
  */
 
-namespace BitFrame\Test\Integration;
+namespace BitFrame\Test\Http;
 
 use PHPUnit\Framework\TestCase;
-use Psr\Http\Message\ServerRequestInterface;
 use BitFrame\Http\ContentNegotiator;
 use BitFrame\Factory\HttpFactory;
 use BitFrame\Parser\MediaParserInterface;
@@ -22,7 +21,76 @@ use BitFrame\Parser\{DefaultMediaParser, JsonMediaParser, XmlMediaParser};
  */
 class ContentNegotiatorTest extends TestCase
 {
+    /**
+     * @runInSeparateProcess
+     */
+    public function testAddMediaParser(): void
+    {
+        $parser = $this->getMockBuilder(MediaParserInterface::class)
+            ->getMock();
 
+        $parser
+            ->method('parse')
+            ->willReturn(['foo' => 'bar']);
+
+        $contentType = ContentNegotiator::CONTENT_TYPE_HTML;
+
+        ContentNegotiator::addMediaParser($contentType, $parser);
+
+        $this->assertInstanceOf(
+            get_class($parser),
+            ContentNegotiator::getMediaParserForContentType($contentType)
+        );
+    }
+
+    /**
+     * @runInSeparateProcess
+     */
+    public function testAddMediaParserByClassName(): void
+    {
+        $parser = $this->getMockBuilder(MediaParserInterface::class)
+            ->getMock();
+
+        $parser
+            ->method('parse')
+            ->willReturn(['foo' => 'bar']);
+
+        $contentType = ContentNegotiator::CONTENT_TYPE_HTML;
+        $parserClass = get_class($parser);
+
+        ContentNegotiator::addMediaParser($contentType, $parserClass);
+
+        $this->assertInstanceOf(
+            $parserClass,
+            ContentNegotiator::getMediaParserForContentType($contentType)
+        );
+    }
+
+    public function mediaParserForContentTypeProvider(): array
+    {
+        return [
+            'html' => [ContentNegotiator::CONTENT_TYPE_HTML, DefaultMediaParser::class],
+            'json' => [ContentNegotiator::CONTENT_TYPE_JSON, JsonMediaParser::class],
+            'xml' => [ContentNegotiator::CONTENT_TYPE_XML, XmlMediaParser::class],
+            'text' => [ContentNegotiator::CONTENT_TYPE_TEXT, DefaultMediaParser::class],
+        ];
+    }
+
+    /**
+     * @dataProvider mediaParserForContentTypeProvider
+     *
+     * @param string $contentType
+     * @param string $parserClassName
+     */
+    public function testGetMediaParserForContentType(
+        string $contentType,
+        string $parserClassName
+    ): void {
+        $this->assertInstanceOf(
+            $parserClassName,
+            ContentNegotiator::getMediaParserForContentType($contentType)
+        );
+    }
 
     public function preferredMediaParserProvider(): array
     {
