+++ test/Http/Message/ResponseDecoratorTest.php
@@ -10,7 +10,7 @@
 
 declare(strict_types=1);
 
-namespace BitFrame\Test\Unit;
+namespace BitFrame\Test\Http\Message;
 
 use ReflectionObject;
 use PHPUnit\Framework\TestCase;
