+++ test/Http/Message/XmlResponseTest.php
@@ -10,49 +10,49 @@
 
 declare(strict_types=1);
 
-namespace BitFrame\Test\Unit;
+namespace BitFrame\Test\Http\Message;
 
 use PHPUnit\Framework\TestCase;
-use BitFrame\Http\Message\HtmlResponse;
+use BitFrame\Http\Message\XmlResponse;
 use TypeError;
 
 /**
- * @covers \BitFrame\Http\Message\HtmlResponse
+ * @covers \BitFrame\Http\Message\XmlResponse
  */
-class HtmlResponseTest extends TestCase
+class XmlResponseTest extends TestCase
 {
-    public function testConstructorAcceptsHtmlString(): void
+    public function testConstructorAcceptsBodyAsString(): void
     {
-        $body = '<html>Lorem ipsum</html>';
-        $response = new HtmlResponse($body);
+        $body = 'Lorem ipsum';
+        $response = new XmlResponse($body);
         $this->assertSame($body, (string) $response->getBody());
         $this->assertSame(200, $response->getStatusCode());
     }
 
     public function testCanAddStatusAndHeader(): void
     {
-        $body = '<html>Not found</html>';
+        $body = '<test>XML</test>';
         $status = 404;
         
-        $response = (new HtmlResponse($body))
+        $response = (new XmlResponse($body))
             ->withStatus($status)
             ->withHeader('x-custom', ['foo-bar']);
         
         $this->assertSame(['foo-bar'], $response->getHeader('x-custom'));
-        $this->assertSame('text/html; charset=utf-8', $response->getHeaderLine('content-type'));
+        $this->assertSame('application/xml; charset=utf-8', $response->getHeaderLine('content-type'));
         $this->assertSame(404, $response->getStatusCode());
         $this->assertSame($body, (string) $response->getBody());
     }
 
     public function testStaticCreateWithCustomContentType(): void
     {
-        $response = HtmlResponse::create('<html>Test</html>')
-            ->withHeader('content-type', 'multipart/form-data');
+        $response = XmlResponse::create('<test>XML</test>')
+            ->withHeader('content-type', 'application/xml-dtd');
         
-        $this->assertSame('multipart/form-data', $response->getHeaderLine('Content-Type'));
+        $this->assertSame('application/xml-dtd', $response->getHeaderLine('Content-Type'));
     }
 
-    public function invalidHtmlContentProvider(): array
+    public function invalidContentProvider(): array
     {
         return [
             'null' => [null],
@@ -68,13 +68,13 @@ class HtmlResponseTest extends TestCase
     }
 
     /**
-     * @dataProvider invalidHtmlContentProvider
+     * @dataProvider invalidContentProvider
      *
      * @param mixed $body
      */
     public function testRaisesExceptionforNonStringContent($body): void
     {
         $this->expectException(TypeError::class);
-        new HtmlResponse($body);
+        new XmlResponse($body);
     }
 }
